python_code/requirement.txt
pandas ==2.0.3
python-dotenv==1.0.1
openai==1.50.2
mlxtend==0.23.1
google-cloud-storage==2.18.2
install pinecone
streamlit
agent_controler